@@ -16,7 +16,6 @@ use Locale;
 use Traversable;
 
 use function array_shift;
-use function get_class;
 use function gettype;
 use function is_array;
 use function is_file;
@@ -126,7 +125,7 @@ class Translator implements TranslatorInterface
             throw new Exception\InvalidArgumentException(sprintf(
                 '%s expects an array or Traversable object; received "%s"',
                 __METHOD__,
-                is_object($options) ? get_class($options) : gettype($options)
+                is_object($options) ? $options::class : gettype($options)
             ));
         }
 
